requirements.txt
Flask==3.0.0
requests==2.31.0
blurhash-python==1.2.2
